requirement.txt
Flask==3.1.2
Flask-Session==0.8.0
requests==2.32.5
python-dotenv==1.1.1
gunicorn==21.2.0
supabase==1.0.4
